requirements.txt
torch
torchaudio
audiocraft
requests
google-api-python-client
google-auth
Pillow
